refactor(Topology): :recycle: Isolate select_services_component
